Moved data files and added docs data - Moved data related files to a data folder making it more understandable - Merged category data files - Added auto doc generation config
