+++ .gitignore
@@ -30,5 +30,7 @@ src/themes.built/
 src/demo/**/*.*
 !src/demo/**/readme.md
 
+data/docs.json
+
 package-lock.json
