+++ data/categories.json
@@ -1,34 +1,42 @@
 [
   {
     "id": "dcb503ac-115e-4ee0-b87c-de67f158d9e7",
-    "name": "Element"
+    "name": "Elements"
   },
   {
     "id": "ab41f517-93e8-4566-ad97-29eb1c6ac944",
-    "name": "Component"
+    "name": "Components"
   },
   {
     "id": "10d95c8e-c1b3-4b95-b4fb-c5d65c177255",
-    "name": "Layout"
+    "name": "Layouts"
   },
   {
     "id": "9baca119-4b64-4877-b8b9-350e3b513947",
-    "name": "Typography"
+    "name": "Typographys"
   },
   {
     "id": "4dabb9b0-1e25-48f8-98e9-336599cc0b37",
-    "name": "Form"
+    "name": "Forms"
   },
   {
     "id": "c4099a33-8125-4b74-8839-0c81aca3aae6",
-    "name": "Table"
+    "name": "Tables"
   },
   {
     "id": "d15da991-148c-4d06-b2df-436a1baf11ac",
-    "name": "Application"
+    "name": "Applications"
   },
   {
     "id": "027ad7d1-afbb-4428-820a-d893a07ee826",
     "name": "Branding"
+  },
+  {
+    "id": "e00a2b4f-e4fc-4165-9d6e-a9760b9c7520",
+    "name": "Mail"
+  },
+  {
+    "id": "e360dda6-15e4-4adb-a1e9-2041df1fd17a",
+    "name": "Charts"
   }
 ]
