+++ data/templates.json
@@ -1,0 +1,7 @@
+[
+  {
+    "id": "92f4852e-1b13-4630-a455-4179f90f9b3d",
+    "name": "main-page",
+    "categories": ["d15da991-148c-4d06-b2df-436a1baf11ac"]
+  }
+]
