--- src/templates/categories.json
@@ -1,22 +1,0 @@
-[
-  {
-    "id": "dcb503ac-115e-4ee0-b87c-de67f158d9e7",
-    "name": "Mail"
-  },
-  {
-    "id": "ab41f517-93e8-4566-ad97-29eb1c6ac944",
-    "name": "Charts"
-  },
-  {
-    "id": "10d95c8e-c1b3-4b95-b4fb-c5d65c177255",
-    "name": "Forms"
-  },
-  {
-    "id": "9baca119-4b64-4877-b8b9-350e3b513947",
-    "name": "Tables"
-  },
-  {
-    "id": "4dabb9b0-1e25-48f8-98e9-336599cc0b37",
-    "name": "Apps"
-  }
-]
