--- src/templates/data.json
@@ -1,12 +1,0 @@
-[
-  {
-    "id": "92f4852e-1b13-4630-a455-4179f90f9b3d",
-    "name": "main-page",
-    "categories": ["4dabb9b0-1e25-48f8-98e9-336599cc0b37"]
-  },
-  {
-    "id": "92f4852e-1b13-4630-a455-4179f90f9b3d",
-    "name": "main-page",
-    "categories": ["4dabb9b0-1e25-48f8-98e9-336599cc0b37"]
-  }
-]
