+++ stencil.config.ts
@@ -6,11 +6,15 @@ import data from './package.json';
 
 export const config: Config = {
   namespace: data.name,
-  outputTargets:[
+  outputTargets: [
     { type: 'dist' },
     {
       type: 'www',
       serviceWorker: null // disable service workers
+    },
+    {
+      type: 'docs-json',
+      file: 'data/docs.json'
     }
   ],
   testing: {
